MCHRF Greedy * Implemented the new greedy MCHRF * Refactor the utilities and splitted in multiple files (facility, town, hazard, utils) * Fixed assign_facility_to_town function * Added docstrings to all functions * Refactor test function into main function with multiple choices
